ID-1421 Move initialisation of guardianClient
