+++ packages/passport/sdk/src/starkEx/passportImxProviderFactory.test.ts
@@ -1,15 +1,15 @@
 import { IMXClient } from '@imtbl/x-client';
 import { ImxApiClients } from '@imtbl/generated-clients';
-import { ConfirmationScreen } from '../confirmation';
 import registerPassportStarkEx from './workflows/registration';
 import { PassportImxProviderFactory } from './passportImxProviderFactory';
 import MagicAdapter from '../magicAdapter';
 import AuthManager from '../authManager';
 import { PassportError, PassportErrorType } from '../errors/passportError';
 import { PassportEventMap } from '../types';
-import { mockUserImx, testConfig } from '../test/mocks';
+import { mockUserImx } from '../test/mocks';
 import TypedEventEmitter from '../utils/typedEventEmitter';
 import { PassportImxProvider } from './passportImxProvider';
+import GuardianClient from '../guardian';
 
 jest.mock('./workflows/registration');
 jest.mock('./passportImxProvider');
@@ -27,17 +27,15 @@ describe('PassportImxProviderFactory', () => {
   const immutableXClient = {
     usersApi: {},
   } as IMXClient;
-  const confirmationScreen = {} as ConfirmationScreen;
-  const config = testConfig;
+  const guardianClient = {} as GuardianClient;
   const passportEventEmitter = new TypedEventEmitter<PassportEventMap>();
   const passportImxProviderFactory = new PassportImxProviderFactory({
-    config,
-    confirmationScreen,
     immutableXClient,
     authManager: mockAuthManager as unknown as AuthManager,
     magicAdapter: mockMagicAdapter as unknown as MagicAdapter,
     passportEventEmitter,
     imxApiClients,
+    guardianClient,
   });
   const mockPassportImxProvider = {};
 
@@ -89,10 +87,9 @@ describe('PassportImxProviderFactory', () => {
         magicAdapter: mockMagicAdapter,
         authManager: mockAuthManager,
         immutableXClient,
-        config,
-        confirmationScreen,
         passportEventEmitter,
         imxApiClients: new ImxApiClients({} as any),
+        guardianClient,
       });
     });
   });
