+++ packages/passport/sdk/src/starkEx/workflows/order.ts
@@ -11,7 +11,7 @@ import {
 import { convertToSignableToken } from '@imtbl/toolkit';
 import { PassportErrorType, withPassportError } from '../../errors/passportError';
 import { UserImx } from '../../types';
-import GuardianClient from '../../guardian/guardian';
+import GuardianClient from '../../guardian';
 
 type CancelOrderParams = {
   request: GetSignableCancelOrderRequest;
