+++ packages/passport/sdk/src/starkEx/workflows/trades.ts
@@ -7,7 +7,7 @@ import {
 } from '@imtbl/core-sdk';
 import { PassportErrorType, withPassportError } from '../../errors/passportError';
 import { UserImx } from '../../types';
-import GuardianClient from '../../guardian/guardian';
+import GuardianClient from '../../guardian';
 
 type CreateTradeParams = {
   request: GetSignableTradeRequest;
