+++ packages/passport/sdk/src/zkEvm/signTypedDataV4.ts
@@ -4,7 +4,7 @@ import {
   Web3Provider,
 } from '@ethersproject/providers';
 import { BigNumber } from 'ethers';
-import GuardianClient from 'guardian/guardian';
+import GuardianClient from 'guardian';
 import { getSignedTypedData } from './walletHelpers';
 import { TypedDataPayload } from './types';
 import { JsonRpcError, RpcErrorCode } from './JsonRpcError';
